transitioning between artist/album/songs now works properly.  fixed a crash when going back from music to home screen
transitioning between artist/album/songs now works properly.  fixed a crash when going back from music to home screen.

i wasn't setting the tag in the transition back to the dashboard.
for album/song transition, pass the artist as well. if we're coming from all artists we need to know this.

Signed-off-by: Eric Friesen <efriesen@gmail.com>