playlist code updates
broke the song and playlist views apart.  they should be apart, they are different.  thanks to java and OOP, it's easy to extend a base class and make this work. yay.
also if you don't have a playlist and long click a song to add to one, it will give you the option to create a new list

Signed-off-by: Eric Friesen <efriesen@gmail.com>