more playlist improvements
removed the "playlist manager" and replaced it with an alert dialog, it looks a lot better and more androidish.  this altered some of the method callbacks i need/don't need, so the old ones have been removed from the base class.
also the playlist is more user friendly for adding/updating.  it's one method, and if there is a valid id, it does an update, if not, it will create a new list

Signed-off-by: Eric Friesen <efriesen@gmail.com>