fix for music button showing album art
i added the resouce id for the on screen buttons so they know which view they are located in.  I needed this for the music button to change the background to show the album art.
just noticed a bug with the notification controller.  when setting the note, the fragment is overwriting the homescreen and driver bars.  need to investigate farther.  for now i'll disable it.