background color choosing
added a color choose to allow for background color choosing.  you can choose two, one top and one bottom for the gradient background color.  also choose an image, change the overall brightness, and reset back to default.

in the works i found another small bug regarding the back button.  i've fixed it in the library and made a pull request to the original author.