start of arduino pin editor
the system will autodetect if you are using a usb device or accessory and setup config accordingly.
once you enter the setup menu, you are shown a picture of the device, you click on the pins and an editor shows up allowing you to configure that bank of pins.
also incldued is improvments to the arduino service, and including the ftdi drivers needed